updating the isAlive
reference 
1. https://github.com/bluespace-ai/bluespace_ai_xsens_ros_mti_driver/pull/17/files
2. https://github.com/bluespace-ai/bluespace_ai_xsens_ros_mti_driver/pull/17

The update of glibc from 2.31 to 2.35 in Ubuntu changed the semantics of pthread_kill. It used to return an error when a dead thread exited, and now it returns 0.

This new behavior is POSIX-compatible as the result of pthread_kill on a thread that died produces undefined behavior, so techincally 0 is valid result.

This isAlive() will return that a dead thread is alive in the newer glibc versions.

This patch removes the use of isAlive and instead uses tryjoin, which should behave correctly for dead threads.